Fix `log` command error when no date is passed in
If `#index` doesn't find a match, it will return `nil`. Correctly handle
this case in `#get_date`.

Since the date can be `nil`, it's easier to do the `Chronic` parsing in
`#get_date` than in `#start`. Finally, move the `#strftime` to `#add` to
allow it to handle `nil` dates at the source.