App#read_config: Clarify the meaning of 'subdomain' when asking for it
The first time I saw it, I wasn't sure what format of subdomain this
prompt wanted, and entered 'acme.harvestapp.com'. Since it's not
documented in the README either, I had to source-dive to find the
answer. Clarify what should be entered here by including an example in
the prompt.