Commands#log: Rewrite to call `task.add`
Copy over the contents of `Commands#start` and replace `task.start` with
`task.add`. This allows us to log the time without actually starting a
timer (and thus having to subsequently stop that timer).

Still need to work out how to eliminate repetition between this and the
`#start` method, but this is now functioning for logging time to any
date (previously it was only possible to `log` to today or yesterday,
otherwise a timer would start).